Codes for generating initial provider list
This file is to generate initial provider list for mothers. This provider list is based on the information in P1_HealthCare_xxx.csv. We don't need to generate provider list every time when we received new P1_HealthCare_xxx file. We just need to run the R_mother_info_processing.r to find new providers which were not recorded in the initial list, and add the new provider name to get a updated list.